requirements.txt
anyio==4.11.0
appdirs==1.4.4
asgiref==3.8.1
baron==0.10.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
cloudinary==1.41.0
colorama==0.4.6
crispy-tailwind==1.0.3
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.2.6
django-cache-url==3.4.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-crispy-forms==2.5
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
marshmallow==4.1.0
packaging==25.0
pillow==12.0.0
psycopg2==2.9.11
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
redbaron==0.9.2
rply==0.7.8
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.2
whitenoise==6.9.0
python-decouple==3.8